+++ ubo_app/services/050-rpi-connect/commands.py
@@ -16,7 +16,6 @@ from ubo_app.store.services.notifications import (
     Importance,
     Notification,
     NotificationDisplayType,
-    NotificationExtraInformation,
     NotificationsAddAction,
 )
 from ubo_app.store.services.rpi_connect import (
@@ -27,6 +26,7 @@ from ubo_app.store.services.rpi_connect import (
     RPiConnectStatus,
     RPiConnectUpdateServiceStateAction,
 )
+from ubo_app.store.services.voice import ReadableInformation
 from ubo_app.utils.apt import is_package_installed
 from ubo_app.utils.async_ import create_task
 from ubo_app.utils.monitor_unit import is_unit_active
@@ -208,7 +208,7 @@ def start_service(is_lightdm_active: bool) -> None:  # noqa: FBT001
                     notification=Notification(
                         title='RPi-Connect',
                         content='LightDM is not running',
-                        extra_information=NotificationExtraInformation(
+                        extra_information=ReadableInformation(
                             text="""\
 LightDM is not running so RPi-Connect will run without screen sharing and you can only \
 use the remote shell feature. To enable screen sharing, start LightDM service in \