+++ ubo_app/services/090-web-ui/reducer.py
@@ -1,8 +1,6 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107
 from __future__ import annotations
 
-import datetime
-import functools
 from dataclasses import replace
 from typing import TYPE_CHECKING
 
@@ -15,15 +13,11 @@ from ubo_app.store.input.types import (
     InputMethod,
     InputResolveAction,
 )
-from ubo_app.store.main import store
 from ubo_app.store.services.notifications import (
-    Notification,
-    NotificationDisplayType,
     NotificationsAction,
-    NotificationsAddAction,
     NotificationsClearByIdAction,
 )
-from ubo_app.store.services.web_ui import WebUIState
+from ubo_app.store.services.web_ui import WebUICheckHotspotEvent, WebUIState
 
 if TYPE_CHECKING:
     from redux import ReducerResult
@@ -34,7 +28,7 @@ DispatchAction = InputCancelAction | NotificationsAction
 def reducer(
     state: WebUIState | None,
     action: InputAction,
-) -> WebUIState | ReducerResult[WebUIState, DispatchAction, None]:
+) -> WebUIState | ReducerResult[WebUIState, DispatchAction, WebUICheckHotspotEvent]:
     if state is None:
         if isinstance(action, InitAction):
             return WebUIState(active_inputs=[])
@@ -49,39 +43,22 @@ def reducer(
                 state,
                 active_inputs=[*state.active_inputs, action.description],
             ),
-            actions=[
-                NotificationsAddAction(
-                    notification=Notification(
-                        id=f'web_ui:pending:{action.description.id}',
-                        icon='󱋆',
-                        title='Web UI',
-                        content=f'[size=18dp]{action.description.prompt}[/size]',
-                        display_type=NotificationDisplayType.STICKY,
-                        is_read=True,
-                        extra_information=action.description.extra_information,
-                        expiration_timestamp=datetime.datetime.now(tz=datetime.UTC),
-                        color='#ffffff',
-                        show_dismiss_action=False,
-                        dismiss_on_close=True,
-                        on_close=functools.partial(
-                            store.dispatch,
-                            InputCancelAction(id=action.description.id),
-                        ),
-                    ),
-                ),
-            ],
+            events=[WebUICheckHotspotEvent(description=action.description)],
         )
     if isinstance(action, InputResolveAction | InputCancelAction):
         return CompleteReducerResult(
             state=replace(
                 state,
-                active_inputs=[
-                    description
-                    for description in state.active_inputs
-                    if description.id != action.id
-                ],
+                active_inputs=(
+                    new_active_inputs := [
+                        description
+                        for description in state.active_inputs
+                        if description.id != action.id
+                    ]
+                ),
             ),
             actions=[NotificationsClearByIdAction(id=f'web_ui:pending:{action.id}')],
+            events=[] if new_active_inputs else [WebUICheckHotspotEvent()],
         )
 
     return state