+++ ubo_app/store/services/web_ui.py
@@ -4,9 +4,17 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from immutable import Immutable
+from redux import BaseEvent
 
 if TYPE_CHECKING:
     from ubo_app.store.input.types import InputDescription
+
+
+class WebUIEvent(BaseEvent): ...
+
+
+class WebUICheckHotspotEvent(WebUIEvent):
+    description: InputDescription | None = None
 
 
 class WebUIState(Immutable):