+++ ubo_app/system/system_manager/reset_button.py
@@ -8,7 +8,7 @@ from ubo_app.logger import get_logger
 logger = get_logger('system-manager')
 
 RESET_BUTTON_PIN = 27
-LONG_PRESS_TIME = 3
+LONG_PRESS_TIME = 5
 SHORT_PRESS_TIME = 0.5
 
 
@@ -26,17 +26,22 @@ def setup_reset_button() -> None:
         release_time = time.time()
         if 'press_time' in state:
             duration = release_time - state['press_time']
-            if duration > LONG_PRESS_TIME:
-                logger.info('Reset button pressed for 3 seconds')
+            logger.info(
+                'Reset button pressed',
+                extra={
+                    'duration': duration,
+                    'SHORT_PRESS_TIME': SHORT_PRESS_TIME,
+                    'LONG_PRESS_TIME': LONG_PRESS_TIME,
+                },
+            )
 
+            if duration > LONG_PRESS_TIME:
                 # Restart the pod
                 subprocess.run(  # noqa: S603
                     ['/usr/bin/env', 'systemctl', 'reboot', '-i'],
                     check=True,
                 )
             elif duration > SHORT_PRESS_TIME:
-                logger.info('Reset button pressed for less than 3 seconds')
-
                 # Kill the UBO process
                 subprocess.run(  # noqa: S603
                     ['/usr/bin/env', 'killall', '-9', 'ubo'],